ML-FastAPI-app/requirements.txt
uvicorn[standard]==0.20.0
gunicorn==20.1.0
fastapi[all]==0.88.0
numpy==1.23.5
opencv-python-headless==4.7.0.72
Pillow==9.5.0
python-multipart==0.0.6
tensorflow==2.12.0